ios 热力图重构

+++ lib/ios/MapView/HeatMapManager.m
@@ -3,7 +3,7 @@
 @interface RCT_EXTERN_MODULE(AMapHeatMapManager, RCTViewManager)
 
 RCT_EXPORT_VIEW_PROPERTY(data, NSArray)
-RCT_EXPORT_VIEW_PROPERTY(radius, double)
+RCT_EXPORT_VIEW_PROPERTY(radius, int)
 RCT_EXPORT_VIEW_PROPERTY(opacity, double)
 
 @end
