+++ lib/ios/Modules/OfflineModule.m
@@ -2,12 +2,12 @@
 
 @interface RCT_EXTERN_MODULE(AMapOffline, NSObject)
 
-RCT_EXTERN_METHOD(download:(NSString)apiKey)
-RCT_EXTERN_METHOD(addListener:(NSString)eventType)
-RCT_EXTERN_METHOD(removeListeners:(NSInteger)count)
+RCT_EXTERN_METHOD(download: (NSString)apiKey)
+RCT_EXTERN_METHOD(addListener: (NSString)eventType)
+RCT_EXTERN_METHOD(removeListeners: (int)count)
 
 - (NSArray<NSString *> *)supportedEvents {
-  return @[@"download"];
+  return @[ @"download" ];
 }
 
 @end
